fix small issues
